# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pokedex API",
        "description": "A simple API for a Pokedex",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
...
        "/validate/pokemon": {
            "get": {
                "summary": "Searches for a random Pokemon in two JSON files",
                "responses": {
                    "200": {
                        "description": "Returns the name and any matching data"
                    },
                    "404": {
                        "description": "Pokemon not found"
                    }
                }
            }
        },
...
    }
}